Move all After() cypress checks to one file
This makes it easier to find checks which run After() each
test, and also easier to turn them on or off en masse.
